+++ types/ranking.ts
@@ -1,0 +1,7 @@
+export interface RankingContextType {
+  examDifficulty: number;
+  assignmentDifficulty: number;
+  interestLevel: number;
+  overallScore: number;
+  overallComment: string;
+}
